Move Request to extension method
